Create TimesSquareParameters component
This component will host a UI for showing and selecting the values of
every parameter. Right now it contains the placeholder list ported from
TimesSquareGitHubPagePanel.
