+++ src/components/TimesSquareGitHubPagePanel/TimesSquareGitHubPagePanel.js
@@ -9,6 +9,7 @@ import Head from 'next/head';
 import Error from 'next/error';
 
 import useTimesSquarePage from '../../hooks/useTimesSquarePage';
+import TimesSquareParameters from '../TimesSquareParameters';
 
 export default function TimesSquareGitHubPagePanel({
   tsPageUrl,
@@ -24,21 +25,7 @@ export default function TimesSquareGitHubPagePanel({
     return <Error statusCode={404} />;
   }
 
-  const { parameters, title, description } = pageData;
-
-  // Merge user-set parameters with defaults
-  const updatedParameters = Object.entries(parameters).map((item) => {
-    if (item[0] in userParameters) {
-      return [item[0], userParameters[item[0]]];
-    } else {
-      return [item[0], item[1].default];
-    }
-  });
-
-  // List items for the parameters
-  const parameterListItems = updatedParameters.map((item) => (
-    <li key={item[0]}>{`${item[0]}: ${item[1]}`}</li>
-  ));
+  const { title, description } = pageData;
 
   return (
     <>
@@ -50,8 +37,10 @@ export default function TimesSquareGitHubPagePanel({
         {description && (
           <div dangerouslySetInnerHTML={{ __html: description.html }}></div>
         )}
-        <p>Notebook parameters:</p>
-        <ul>{parameterListItems}</ul>
+        <TimesSquareParameters
+          pageData={pageData}
+          userParameters={userParameters}
+        />
       </div>
     </>
   );
