+++ src/components/TimesSquareParameters/index.js
@@ -1,0 +1,2 @@
+export * from './TimesSquareParameters';
+export { default } from './TimesSquareParameters';
